Create KHAN A. PROJECT
Here I created a database for a shoe store with the prices and options reflecting what I would like to see in stores. I then selected them by price to show what options were more "expensive" and then which shoe option was the most expensive. This was mostly a silly project just to start practicing SQL.